Add Display

+++ src/cmp/Battery.h
@@ -12,7 +12,7 @@ public:
 	~Battery() = default;
 
 	virtual Core::CoreStatus Initialize( void ) final override;
-	virtual void Update( const uint32_t currentTime ) final override;
+	virtual void Update( const uint64_t currentTime ) final override;
 
 	virtual BatteryState GetState( void ) final override;
 	virtual uint16_t     GetVoltage( void ) final override;
