+++ src/cmp/Mpu9150Interface.h
@@ -16,7 +16,7 @@ public:
 	~Mpu9150Interface()      = default;
 
 	virtual Core::CoreStatus Initialize( void )       = 0;
-	virtual void Update( const uint32_t currentTime ) = 0;
+	virtual void Update( const uint64_t currentTime ) = 0;
 
 	virtual Vector3 ReadAcc( void )  = 0;
 	virtual Vector3 ReadGyr( void )  = 0;
