+++ src/cmp/SoftwareInterface.h
@@ -15,14 +15,14 @@ public:
 	~SoftwareInterface() = default;
 
 	virtual Core::CoreStatus Initialize( void )       = 0;
-	virtual void Update( const uint32_t currentTime ) = 0;
+	virtual void Update( const uint64_t currentTime ) = 0;
 
 	virtual Version GetVersion( void )             = 0;
-	virtual uint32_t GetMinTime( void )            = 0;
-	virtual uint32_t GetMaxTime( void )            = 0;
+	virtual uint64_t GetMinTime( void )            = 0;
+	virtual uint64_t GetMaxTime( void )            = 0;
 	virtual void ResetTiming( void )               = 0;
-	virtual void SetMinTime( const uint32_t time ) = 0;
-	virtual void SetMaxTime( const uint32_t time ) = 0;
+	virtual void SetMinTime( const uint64_t time ) = 0;
+	virtual void SetMaxTime( const uint64_t time ) = 0;
 
 private:
 };
