+++ src/drv/TickInterface.h
@@ -9,8 +9,8 @@ public:
 	~TickInterface() = default;
 
 	virtual uint64_t GetUs( void )           = 0;
-	virtual uint32_t GetMs( void )           = 0;
-	virtual void DelayMs( uint16_t delayMs ) = 0;
-	virtual void DelayUs( uint32_t delayUs ) = 0;
+	virtual uint64_t GetMs( void )           = 0;
+	virtual void DelayMs( uint64_t delayMs ) = 0;
+	virtual void DelayUs( uint64_t delayUs ) = 0;
 };
 }
