+++ src/drv/UartInterface.h
@@ -10,7 +10,7 @@ public:
 	~UartInterface() = default;
 
 	virtual Core::CoreStatus Initialize( void )       = 0;
-	virtual void Update( const uint32_t currentTime ) = 0;
+	virtual void Update( const uint64_t currentTime ) = 0;
 
 	virtual void Send( const char *data, const size_t len )    = 0;
 	virtual void Send( const uint8_t *data, const size_t len ) = 0;
