+++ src/srv/ServiceControl.cpp
@@ -12,7 +12,7 @@ Core::CoreStatus ServiceControl::Initialize ( void )
 	return ( this->mClusterServo.Initialize() );
 }
 
-void ServiceControl::Update ( const uint32_t currentTime )
+void ServiceControl::Update ( const uint64_t currentTime )
 {
 	this->mClusterServo.Update( currentTime );
 	this->mClusterServo.GetComponent().GetPca9685( this->mStepPca9685 ).Update( currentTime );
