+++ src/srv/ServiceProximity.h
@@ -11,7 +11,7 @@ public:
 	~ServiceProximity() = default;
 
 	virtual Core::CoreStatus Initialize( void ) final override;
-	virtual void Update( const uint32_t currentTime ) final override;
+	virtual void Update( const uint64_t currentTime ) final override;
 
 protected:
 	ClusterProximity &mClusterProximity;
