+++ unittests/mock/cmp/MockLed.h
@@ -6,7 +6,7 @@ namespace Component {
 class MockLed : public Component::LedInterface {
 public:
 	MOCK_METHOD0(Initialize, Core::CoreStatus(void) );
-	MOCK_METHOD1(Update, void( const uint32_t ) );
+	MOCK_METHOD1(Update, void( const uint64_t ) );
 	MOCK_METHOD0(On, Core::CoreStatus(void) );
 	MOCK_METHOD0(Off, Core::CoreStatus(void) );
 	MOCK_METHOD0(Toggle, Core::CoreStatus(void) );
